553: Extract campaign refresh button
